% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcBins.R
\name{computeVpcBins}
\alias{computeVpcBins}
\title{Generate Bins}
\usage{
computeVpcBins(
  data,
  split = NULL,
  type = "continuous",
  binsSettings = NULL,
  is.binsName = FALSE,
  dataName = "obs"
)
}
\arguments{
\item{data}{(\emph{vector(double)}) vector of data (times or obs) to compute bins.}

\item{split}{(\emph{vector}) split category associated with data (no split by default).}

\item{type}{in {"continuous", "categorical"} type of data ("continuous" by default).}

\item{binsSettings}{(\emph{binSettingsClass}) (\emph{optional}) Bins settings.}

\item{is.binsName}{(\emph{bool}) (\emph{optional}) if TRUE add a column binsName in P(x<dataName<y) (default FALSE).}

\item{dataName}{(\emph{string}) (\emph{optional}) dataName when is.binsName is TRUE.}
}
\value{
bins dataframe
}
\description{
Generate Bins
}
\examples{
\dontrun{
  computeVpcBins(seq_len(100), binsSettings = getBinsSettings(nbBins = 3))
}
}
