% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcTheme.R
\name{continuous_theme}
\alias{continuous_theme}
\title{Get default VPC theme for continuous data}
\usage{
continuous_theme()
}
\value{
A list with vpc theme specifiers
}
\description{
Get default VPC theme for continuous data
}
\examples{
theme <- continuous_theme()
}
\seealso{
\link{createVpcTheme}, \link{discrete_theme}, \link{event_theme}
}
