% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateModelSelection.R
\name{covariateModelSelection}
\alias{covariateModelSelection}
\title{Get simulated predictions}
\usage{
covariateModelSelection(penalization = "BIC", nb.model = 1,
  trans.cov = "all", param.cov = "all")
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the simulated individual parameters :
}
\examples{
\dontrun{
r = getSimulatedPredictions()
names(r)
    "Cc"  "E" 
}
}
