% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcTheme.R
\name{createVpcTheme}
\alias{createVpcTheme}
\alias{continuous_theme}
\alias{discrete_theme}
\alias{event_theme}
\title{Create a customized VPC theme}
\usage{
createVpcTheme(update = NULL)

continuous_theme()

discrete_theme()

event_theme()
}
\arguments{
\item{update}{list containing the plot elements to be updated. Run `create_vpc_theme()` with no arguments to show an overview of default plot elements.}
}
\value{
A list with vpc theme specifiers
}
\description{
Create a customized VPC theme

`continuous_theme()` returns a default VPC theme for continuous_theme data

`discrete_theme()` returns a default VPC theme for discrete data

`event_theme()` returns a default VPC theme for event data
}
\details{
This function creates a theme that customizes how the VPC looks, i.e. legend, colors, fills, transparencies, linetypes an sizes, etc. The following arguments can be specified in the input list:
\itemize{
 \item{obs_color} color for observations points
 \item{obs_size} size for observation points
 \item{obs_shape} shape for observation points
 \item{obs_legend} legend name for observation points (if NA, no legend for obs)
 \item{cens_color} color for censored points
 \item{cens_shape} shape for censored points
 \item{cens_size} size for censored points
 \item{cens_legend} legend name for censored points (if NA, no legend for cens)
 \item{emp_perc_color} color for empirical percentiles
 \item{emp_perc_linetype} linetype for empirical percentiles
 \item{emp_perc_size} size for empirical percentiles
 \item{emp_perc_legend} legend name for empirical percentiles
 \item{theo_perc_color} color for theorical percentiles
 \item{theo_perc_linetype} linetype for theorical percentiles
 \item{theo_perc_size} size for theorical percentiles
 \item{theo_perc_legend} legend name for theorical percentiles
 \item{theo_pi_median_color} color for prediction interval median
 \item{theo_pi_median_fill} fill for prediction interval median
 \item{theo_pi_median_alpha} alpha for prediction interval median
 \item{theo_pi_median_legend} legend name for prediction interval median
 \item{theo_pi_perc_color} color for prediction interval percentiles
 \item{theo_pi_perc_fill} fill for prediction interval percentiles
 \item{theo_pi_perc_alpha} alpha for prediction interval percentiles
 \item{theo_pi_perc_legend} legend name for prediction interval percentiles
 \item{outlier_areas_color} color for outlier areas
 \item{outlier_areas_fill} fill for outlier areas
 \item{outlier_areas_perc_alpha} alpha for outlier areas
 \item{outlier_areas_perc_legend} legend name for outlier areas
 \item{outlier_dots_color} color for outliers dots
 \item{outlier_dots_size} size for outliers dots
 \item{outlier_dots_shape} shape for outliers dots
 \item{outlier_dots_legend} legend name for outliers dots 
 \item{bins_color} color for bin limits
 \item{bins_linetype} linetype for bin limits
 \item{bins_size} size for bin limits
 \item{bins_legend} legend name for bin limits
}
}
\examples{
theme <- new_vpc_theme(update = list(
  obs_color = "red",
  cens_color = rgb(70, 130, 180, maxColorValue = 255),
  theo_pi_median_alpha = 0.3
))
}
