% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{cutData}
\alias{cutData}
\title{Get dataset to get only observations after the last dose}
\usage{
cutData(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
observations after last dose (dataframe)
}
\description{
Get dataset to get only observations after the last dose
}
\examples{
\dontrun{
cutData()
}
}
