% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcTheme.R
\name{discrete_theme}
\alias{discrete_theme}
\title{Get default VPC theme for discrete data}
\usage{
discrete_theme()
}
\value{
A list with vpc theme specifiers
}
\description{
Get default VPC theme for discrete data
}
\examples{
theme <- continuous_theme()
}
\seealso{
\link{createVpcTheme}, \link{continuous_theme}, \link{event_theme}
}
