% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcTheme.R
\name{event_theme}
\alias{event_theme}
\title{Get default VPC theme for event data}
\usage{
event_theme()
}
\value{
A list with vpc theme specifiers
}
\description{
Get default VPC theme for event data
}
\examples{
theme <- continuous_theme()
}
\seealso{
\link{createVpcTheme}, \link{discrete_theme}, \link{continuous_theme}
}
