% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcStats.R
\name{getBinsSettings}
\alias{getBinsSettings}
\title{Bins settings}
\usage{
getBinsSettings(
  is.fixedBins = FALSE,
  fixedBins = c(),
  criteria = "leastsquare",
  is.fixedNbBins = FALSE,
  nbBins = 10,
  binRange = c(5, 20),
  nbBinData = c(10, 200),
  nbDataPoints = 100
)
}
\arguments{
\item{is.fixedBins}{emph{[Optional]} (boolean) If TRUE, specify manually bin list (default FALSE).}

\item{fixedBins}{emph{[Optional]} (list[integer]) Define mannually a list of bins. To use when `useFixedBins` is set to TRUE.}

\item{criteria}{emph{[Optional]} Choose the bining criteria among `equalwidth`, `equalsize` or `leastsquare` (default leastsquare).}

\item{is.fixedNbBins}{emph{[Optional]} (boolean) If TRUE define a fixed number of bins, else define a range for automatic selection (default FALSE).}

\item{nbBins}{emph{[Optional]} (integer) Define a fixed number of bins (default 10).}

\item{binRange}{emph{[Optional]} (list[integer]) Define a range for the number of bins (default c(5, 20)).}

\item{nbBinData}{emph{[Optional]} (list[integer]) Define a range for the number of data points per bin (default c(10, 200)).}

\item{nbDataPoints}{emph{[Optional]}[event data] (integer) Number of data point in event time grid (default 100)}
}
\value{
binsSettingClass object
}
\description{
Bins settings
}
\examples{
\dontrun{
  getBinsSettings(is.fixedBins = TRUE, fixedBins = c(1, 2, 3))
}
}
