% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{getCensoringInformation}
\alias{getCensoringInformation}
\title{Get censoring information from input dataset}
\usage{
getCensoringInformation()
}
\value{
Censoring columns of input dataset (dataframe).
}
\description{
Get censoring information from input dataset
}
\examples{
\dontrun{
getCensoringInformation()
}
}
