% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{getDoseInformation}
\alias{getDoseInformation}
\title{Get dose rows from input dataset}
\usage{
getDoseInformation()
}
\value{
Dose rows of input dataset (dataframe)
}
\description{
Get dose rows from input dataset
}
\examples{
\dontrun{
getDoseInformation()
}
}
