% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedCovarianceMatrix}
\alias{getEstimatedCovarianceMatrix}
\title{Get estimated covariance and correlation matrices}
\usage{
getEstimatedCovarianceMatrix()
}
\value{
a list of two matrices.
}
\description{
Get estimated covariance and correlation matrices
}
\examples{
\dontrun{
# Assume that the Monolix project "warfarinPKPD.mlxtran" has been loaded
r = getEstimatedCovarianceMatrix()  # r is a list with elements "cor.matrix" and "cov.matrix"

# See http://monolix.lixoft.com/rsmlx/newconnectors/ for more detailed examples
# Download the demo examples here: http://monolix.lixoft.com/rsmlx/installation
}
}
