% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedIndividualParameters2}
\alias{getEstimatedIndividualParameters2}
\title{Get estimated individual and population parameters}
\usage{
getEstimatedIndividualParameters2()
}
\value{
a list of data frames.
}
\description{
Get the individual individual parameters, the population parameters with the population covariates 
and the population parameters with the individual covariates.
}
\examples{
\dontrun{
r = getEstimatedIndividualParameters2()
names(r)
    "saem"  "conditionalMean" "conditionalSD"   "conditionalMode" "popPopCov"    "popIndCov" 
}
}
