% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedPredictions}
\alias{getEstimatedPredictions}
\title{Get estimated predictions}
\usage{
getEstimatedPredictions()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the estimated individual parameters :
}
\examples{
\dontrun{
r = getEstimatedPredictions()
names(r)
    "Cc"  "E" 
}
}
