% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedResiduals}
\alias{getEstimatedResiduals}
\title{Get estimated residuals}
\usage{
getEstimatedResiduals()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the residuals computed from the individual predictions obtained 
with the estimated individual parameters:
}
\examples{
\dontrun{
r = getEstimatedResiduals()
names(r)
    "y1"  "y2" 
}
}
