% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{getPKObservations}
\alias{getPKObservations}
\title{Get observations rows from input dataset}
\usage{
getPKObservations()
}
\value{
PK Observations rows of input dataset (dataframe)
}
\description{
Get observations rows from input dataset
}
\examples{
\dontrun{
getPKObservations()
}
}
