% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getSimulatedPredictions}
\alias{getSimulatedPredictions}
\title{Get simulated predictions}
\usage{
getSimulatedPredictions()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the simulated individual parameters :
}
\examples{
\dontrun{
r = getSimulatedPredictions()
names(r)
    "Cc"  "E" 
}
}
