% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getSimulatedResiduals}
\alias{getSimulatedResiduals}
\title{Get simulated residuals}
\usage{
getSimulatedResiduals()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the residuals computed from the individual predictions obtained 
with the simulated individual parameters:
}
\examples{
\dontrun{
r = getSimulatedResiduals()
names(r)
    "y1"  "y2" 
}
}
