% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmlx-init.R
\name{initRsmlx}
\alias{initRsmlx}
\title{Initialize Rsmlx library}
\usage{
initRsmlx(path = NULL, errors = TRUE, warnings = TRUE, info = FALSE)
}
\arguments{
\item{path}{(\emph{optional}) path to the installation directory of the Lixoft suite
If no path is given, the one written in the <user home>/lixoft/lixoft.ini file is used}

\item{errors}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors errors (default TRUE)}

\item{warnings}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors warnings (default TRUE)}

\item{info}{(\emph{optional}) (\emph{boolean}) Display/Hide lixoft connectors info (default FALSE)}
}
\value{
A list:
\itemize{
  \item \code{software}: the software that is used (should be monolix with Rsmlx)
  \item \code{path}: the path to MonolixSuite
  \item \code{version}: the version of MonolixSuite that is used
  \item \code{status}: boolean equaling TRUE if the initialization has been successful.
}
}
\description{
Initialize Rsmlx library
}
\examples{
\dontrun{
initRsmlx()  # print the info about Monolix and lixoftConnectors
initRsmlx(path="C:/ProgramData/Lixoft/MonolixSuite2019R1")  # use MonolixSuite 2019R1
}
}
