% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{readmlxDataset}
\alias{readmlxDataset}
\title{Read pkanalix input dataset}
\usage{
readmlxDataset(obsid = NULL)
}
\value{
Input dataset in dataframe format. Headers are renamed to match mlx names
}
\description{
Read pkanalix input dataset
}
\examples{
\dontrun{
readmlxDataset()
}
}
