% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resMonolix}
\alias{resMonolix}
\title{Monolix results}
\format{
A R list
}
\source{
Monolix demos
}
\usage{
resMonolix
}
\description{
Monolix results used by the Rsmlx examples
}
\references{
Rsmlx website: http://rsmlx.webpopix.org
}
\keyword{datasets}
