% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RsmlxDemo1.project}
\alias{RsmlxDemo1.project}
\title{Monolix project for warfarin PK - 1}
\format{
A text file
}
\source{
Monolix project
}
\usage{
RsmlxDemo1.project
}
\description{
RsmlxDemo1.mlxtran is a Monolix project for modelling the pharmacokinetics (PK) of warfarin 
using a PK model with parameters ka, V, Cl. There is no covariate in the model.
}
\references{
Rsmlx doumentation
}
\keyword{datasets}
