\name{RsmlxDemo2.project}
\docType{data}
\alias{RsmlxDemo2.project}
\title{Monolix project for warfarin PK - 2}
\description{
  RsmlxDemo2.mlxtran is a Monolix project for modelling the pharmacokinetics (PK) of warfarin 
using a PK model with parameters ka, V, Cl. Here, V and Cl are function of weight.
}
\usage{RsmlxDemo2.project}
\format{A text file}
\source{Monolix project}
\references{Rsmlx doumentation}
\keyword{datasets}

