% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warfarin.data}
\alias{warfarin.data}
\title{warfarin PKPD data}
\format{
A csv file
}
\source{
Monolix demos
}
\usage{
warfarin.data
}
\description{
The warfarin PK and PD data for 32 patients
}
\references{
O’Reilly (1968). Studies on coumarin anticoagulant drugs. Initiation of warfarin therapy without a loading dose. Circulation 1968, 38:169-177.
}
\keyword{datasets}
