% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichPKmodel.R
\name{whichPKmodel}
\alias{whichPKmodel}
\title{Find a Monolix PK model}
\usage{
whichPKmodel(parameter, mlxPath = NULL, pkPath = NULL)
}
\arguments{
\item{parameter}{a vector of PK parameter names}

\item{mlxPath}{path to Monolix install}

\item{pkPath}{path to the Monolix PK library}
}
\description{
Return the path of the Monolix PK model defined by a list of parameter names
See http://rsmlx.webpopix.org/whichPKmodel/ for more details.
}
\examples{
\dontrun{
whichPKmodel(parameter=c("Tlag", "Tk0", "V", "Cl"))
}
}
