% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDataSmlx.R
\name{writeDataSmlx}
\alias{writeDataSmlx}
\title{Write Simulx Dataset}
\usage{
writeDataSmlx(
  filename = "simulated_dataset.csv",
  sep = ",",
  ext = "csv",
  nbdigits = 5,
  mapObservation = NULL
)
}
\arguments{
\item{filename}{(\emph{string}) (\emph{optional}) file path to dataset.
(default "simulated_dataset.csv")
In case of multiple replicates, the function creates one dataset per replicate with name $filename_repi
If filename contains an extension, it must be "csv" or "txt". If it does not, extension is defined by `ext` argument.}

\item{sep}{(\emph{string}) (\emph{optional}) Separator used to write dataset file. (default ",")
It must be one of "\\t", " ", ";", ","}

\item{ext}{(\emph{bool}) (\emph{optional}) Extension used to write dataset file. (default "csv")
It must be one of "csv", "txt"
To defined only if filename with no extension}

\item{nbdigits}{(\emph{integer}) (\emph{optional}) number of decimal digits in output file.
(default = 5)}

\item{mapObservation}{(\emph{name vector}) (\emph{optional}) mapping of observation name}
}
\value{
a dataframe if one single simulation, a list of dataframe if multiple replicates.
}
\description{
Format outputs of simulx simulations and write datasets in monolix
and pkanalix project format.
}
\details{
WARNING: `writeData` function is not implemented for simulx project with regressors in MonolixSuite version 2020R1
}
