% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootmlx}
\alias{bootmlx}
\title{Non parametric bootstrapping}
\usage{
bootmlx(project, nboot = 100, dataFolder = NULL, settings = NULL)
}
\arguments{
\item{project}{Monolix project}

\item{nboot}{number of bootstrat replicates (default= 100)}

\item{dataFolder}{[optional] folder where the resampled datasets are stored}

\item{settings}{a list of settings for the resampling and the results:
\itemize{
\item \code{N} the number of individuals in each bootstrap data set 
(default value is the  number of individuals in the original data set).
\item \code{newResampling} boolean to resample or not the data sets (default=FALSE).
\item \code{strata} categorical covariate of the project. The original distribution of this covariate
is maintained in each resampled data set if strata is defined (default=NULL).
\item \code{plot} boolean to choose if the distribution of the bootstraped esimates is displayed
(default = TRUE)
\item \code{level} level of the bootstrap confidence intervals of the population parameters
(default = 0.90)
}}
}
\value{
a data frame with the bootstrap estimates
}
\description{
Generate replicates of the original data using random sampling with replacement.
Population parameters are then estimated from each replicate.
}
