% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootmlx}
\alias{bootmlx}
\title{Bootstrapping - case resampling}
\usage{
bootmlx(project, nboot = 100, dataFolder = NULL, settings = NULL)
}
\arguments{
\item{project}{Monolix project}

\item{nboot}{[optional] number of bootstrap replicates (default=100)}

\item{dataFolder}{[optional] folder where already generated datasets are stored, e.g dataFolder="./dummy_project/boot/" (default: data set are generated by bootmlx)}

\item{settings}{[optional] a list of settings for the resampling and the results:
\itemize{
\item \code{N} the number of individuals in each bootstrap data set 
(default value is the  number of individuals in the original data set).
\item \code{newResampling} boolean to generate the data sets again if they already exist (default=FALSE).
\item \code{covStrat} a categorical covariate of the project. The original distribution of this covariate
is maintained in each resampled data set if covStrat is defined (default=NULL). Notice that if the categorical covariate is varying 
within the subject (in case of IOV), it will not be taken into account.
\item \code{plot} boolean to choose if the distribution of the bootstraped esimates is displayed
(default = TRUE)
\item \code{level} level of the bootstrap confidence intervals of the population parameters
(default = 0.90)
}}
}
\value{
a data frame with the bootstrap estimates
}
\description{
Generate replicates of the original data using random sampling with replacement.
Population parameters are then estimated from each replicate.
}
