% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmlx.R
\name{buildmlx}
\alias{buildmlx}
\title{Automatic model building}
\usage{
buildmlx(project, final.project = NULL, model = c("residualError",
  "covariate", "correlation"), penalization = "BIC", max.iter = 20,
  covToTransform = "none", paramToUse = "all", linearization = FALSE,
  seqcc = TRUE, nb.model = 1, print = TRUE, direction = "full",
  steps = 1000, p.min = 1, lambda = "cv", settings = NULL)
}
\arguments{
\item{project}{a string: the initial Monolix project}

\item{final.project}{a string: the final Monolix project (default adds "_built" to the original project)}

\item{model}{components of the model to optimize, default=c("residualError", "covariate", "correlation")}

\item{penalization}{penalization criteria to optimize c({"BIC"}, "AIC", "lasso")}

\item{max.iter}{maximum number of iterations (default=20)}

\item{covToTransform}{list of (oontinuous) covariates to be log-transformed (default="none")}

\item{paramToUse}{list of parameters possibly function of covariates (default="all")}

\item{linearization}{TRUE/{FALSE} whether the computation of the likelihood is based on a linearization of the model (default=FALSE)}

\item{seqcc}{{TRUE}/FALSE whether the covariate model is built before the correlation model (default=TRUE)}

\item{nb.model}{number of models to display at each iteration (default=1)}

\item{print}{{TRUE}/FALSE display the results (default=TRUE)}

\item{direction}{method for covariate search c({"full"}, "both", "backward", "forward"), (default="full")}

\item{steps}{maximum number of iteration for stepAIC/BIC (default=1000)}

\item{p.min}{minimum p-value (for the correlation test) for keeping a covariate in a model  (default=1)}

\item{lambda}{penalization coefficient for "lasso" (default="cv"). Either "cv" (cross-validation) or a positive real value (default="cv")}

\item{settings}{when penalization="lasso", settings used for package glmnet}
}
\value{
a new Monolix project with a new statistical model.
}
\description{
Iterative procedure to accelerate and optimize the process of model building by identifying 
at each step how best to improve some of the model components. This method allows to find 
the optimal statistical model which minimizes some information criteria in very few steps.
}
\examples{
\dontrun{
r = buildmlx("warfPK.mlxtran")
}
}
