% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmlx.R
\name{buildmlx}
\alias{buildmlx}
\title{Automatic model building}
\usage{
buildmlx(project, final.project = NULL, model = c("residualError",
  "covariate", "correlation"), penalization = "BIC", max.iter = 20,
  trans.cov = "all", param.cov = "all", seqcc = TRUE, nb.model = 1,
  print = TRUE)
}
\arguments{
\item{project}{a string: the initial Monolix project}

\item{final.project}{a string: the final Monolix project (default is the original project)}

\item{model}{component of the model to optimize, default=c("residualError", "covariate", "correlation")}

\item{penalization}{penalization criteria to optimize c({"BIC"}, "AIC", "lasso")}

\item{max.iter}{maximum number of iterations (default=20)}

\item{trans.cov}{list of covariates to be log-transformed (default="all")}

\item{param.cov}{list of parameters which could be function of covariates (default="all")}

\item{seqcc}{{TRUE}/FALSE whether the covariate model is built before the correlation model (default=TRUE)}

\item{nb.model}{number of models to display at each iteration (default=1)}

\item{print}{{TRUE}/FALSE display the results (default=TRUE)}
}
\value{
a new Monolix project with a new statistical model.
}
\description{
Iterative procedure to accelerate and optimize the process of model building by identifying 
at each step how best to improve some of the model components. This method allows to find 
the optimal statistical model which minimizes some information criteria in very few steps.
}
\examples{
\dontrun{
r = buildmlx("warfPK.mlxtran")
}
}
