% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmlx.R
\name{buildmlx}
\alias{buildmlx}
\title{Automatic model building}
\usage{
buildmlx(project, final.project = NULL, model = "all", paramToUse = "all",
  covToTest = "all", covToTransform = "none", criterion = "BIC",
  direction = NULL, max.iter = 20, print = TRUE, nb.model = 1,
  linearization = FALSE, seqcc = FALSE, p.max = 1, steps = 1000,
  exp.iter = 2)
}
\arguments{
\item{project}{a string: the initial Monolix project}

\item{final.project}{a string: the final Monolix project (default adds "_built" to the original project)}

\item{model}{components of the model to optimize c("residualError", "covariate", "correlation"), (default="all")}

\item{paramToUse}{list of parameters possibly function of covariates (default="all")}

\item{covToTest}{components of the covariate model that can be modified   (default="all")}

\item{covToTransform}{list of (continuous) covariates to be log-transformed (default="none")}

\item{criterion}{penalization criterion to optimize c({"BIC"}, "AIC", gamma)}

\item{direction}{method for covariate search c({"full"}, "both", "backward", "forward"), (default="full" or "both")}

\item{max.iter}{maximum number of iterations (default=20)}

\item{print}{{TRUE}/FALSE display the results (default=TRUE)}

\item{nb.model}{number of models to display at each iteration (default=1)}

\item{linearization}{TRUE/{FALSE} whether the computation of the likelihood is based on a linearization of the model (default=FALSE)}

\item{seqcc}{TRUE/{FALSE} whether the covariate model is built before the correlation model (default=FALSE)}

\item{p.max}{maximum p-value (for the correlation test) for keeping a covariate in a model  (default=1)}

\item{steps}{maximum number of iteration for stepAIC (default=1000)}

\item{exp.iter}{number of iterations during the exploratory phase (default=1)}
}
\value{
a new Monolix project with a new statistical model.
}
\description{
buildmlx uses SAMBA (Stochastic Approximation for Model Building Algorithm), an iterative procedure to accelerate and optimize the process of model building by identifying 
at each step how best to improve some of the model components. This method allows to find 
the optimal statistical model which minimizes some information criterion in very few steps.
}
\details{
Penalization criterion can be either a custom penalization of the form gamma*(number of parameters),
AIC (gamma=2) or BIC (gamma=log(N)).

Several strategies can be used for building the covariate model at each iteration of the algorithm: 
\code{direction="full"} means that all the possible models are compared (default when the number of covariates
is less than 10). Othrwise, \code{direction} is the mode of stepwise search of \code{stepAIC {MASS}}, 
can be one of "both", "backward", or "forward", with a default of "both" when there are at least 10 covariates.

See http://rsmlx.webpopix.org for more details.
}
\examples{
\dontrun{
r = buildmlx("warfPK.mlxtran")
}
}
