% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmlx.R
\name{buildmlx}
\alias{buildmlx}
\title{Automatic model building}
\usage{
buildmlx(project, final.project = NULL, model = c("residualError",
  "covariate", "correlation"), paramToUse = "all", covToUse = "all",
  covToTransform = "none", penalization = "BIC", direction = "full",
  max.iter = 20, print = TRUE, nb.model = 1, linearization = FALSE,
  seqcc = TRUE, p.min = 1, steps = 1000, lambda = "cv",
  glmnet.settings = NULL)
}
\arguments{
\item{project}{a string: the initial Monolix project}

\item{final.project}{a string: the final Monolix project (default adds "_built" to the original project)}

\item{model}{components of the model to optimize, default=c("residualError", "covariate", "correlation")}

\item{paramToUse}{list of parameters possibly function of covariates (default="all")}

\item{covToUse}{components of the covariate model that can be modified   (default="all")}

\item{covToTransform}{list of (continuous) covariates to be log-transformed (default="none")}

\item{penalization}{penalization criteria to optimize c({"BIC"}, "AIC", "lasso")}

\item{direction}{method for covariate search c({"full"}, "both", "backward", "forward"), (default="full")}

\item{max.iter}{maximum number of iterations (default=20)}

\item{print}{{TRUE}/FALSE display the results (default=TRUE)}

\item{nb.model}{number of models to display at each iteration (default=1)}

\item{linearization}{TRUE/{FALSE} whether the computation of the likelihood is based on a linearization of the model (default=FALSE)}

\item{seqcc}{{TRUE}/FALSE whether the covariate model is built before the correlation model (default=TRUE)}

\item{p.min}{minimum p-value (for the correlation test) for keeping a covariate in a model  (default=1)}

\item{steps}{maximum number of iteration for stepAIC/BIC (default=1000)}

\item{lambda}{penalization coefficient for "lasso". Either "cv" (cross-validation) or a positive real value (default="cv")}

\item{glmnet.settings}{when penalization="lasso", list of settings used for package glmnet}
}
\value{
a new Monolix project with a new statistical model.
}
\description{
Iterative procedure to accelerate and optimize the process of model building by identifying 
at each step how best to improve some of the model components. This method allows to find 
the optimal statistical model which minimizes some information criteria in very few steps.
}
\examples{
\dontrun{
r = buildmlx("warfPK.mlxtran")
}
}
