% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateSearch.R
\name{covariateSearch}
\alias{covariateSearch}
\title{Covariate model building}
\usage{
covariateSearch(project, final.project = NULL, covToTest = NULL,
  paramToUse = NULL, settings = NULL)
}
\arguments{
\item{project}{a Monolix project}

\item{final.project}{a string: the final Monolix project (default adds method name to original project name)}

\item{covToTest}{[optional] list of covariates to test. By default, all covariates are tested}

\item{paramToUse}{[optional] list of parameters which may be function of covariates. By default, all parameters are used.}

\item{settings}{[optional] defines the settings search. This is a list of all settings for the search. 
The settings are pInclusion, pElimination, linearization, rankedSCM, criteria, updateInit. By default, the values are .1, .05, FALSE, TRUE, 'LRT' and TRUE.}
}
\description{
Automatic search of the best covariate model. 
Several methods for covariate model building are proposed
\itemize{
\item COSSAC: ranked SCM
\item SCM: SCM method
}
}
