% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{GetEstimatedCovarianceMatrix}
\alias{GetEstimatedCovarianceMatrix}
\title{Get estimated covariance and correlation matrices}
\usage{
GetEstimatedCovarianceMatrix()
}
\value{
a list of two matrices.
}
\description{
Get estimated covariance and correlation matrices
}
\examples{
\dontrun{
r = GetEstimatedCovarianceMatrix()
names(r)
    "cor.matrix" "cov.matrix"
}
}
