% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedIndividualParameters2}
\alias{getEstimatedIndividualParameters2}
\title{Get estimated individual and population parameters}
\usage{
getEstimatedIndividualParameters2()
}
\value{
a list of data frames.
}
\description{
Get the individual individual parameters, the population parameters with the population covariates 
and the population parameters with the individual covariates.
}
\examples{
\dontrun{
# Assume that the Monolix project "warfarinPKPD.mlxtran" has been loaded
r = getEstimatedIndividualParameters2() 

# r is a list with elements "saem", "conditionalMean", "conditionalSD",   "conditionalMode",
# "popPopCov" and "popIndCov"

# See http://rsmlx.webpopix.org/userguide/newconnectors/ for more detailed examples
# Download the demo examples here: http://rsmlx.webpopix.org/installation
}
}
