% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedPredictions}
\alias{getEstimatedPredictions}
\title{Get estimated predictions}
\usage{
getEstimatedPredictions()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the estimated individual parameters :
}
\examples{
\dontrun{
# Assume that the Monolix project "warfarinPKPD.mlxtran" has been loaded
r = getEstimatedPredictions() # r is a list with elements "y1" and "y2"

# See http://rsmlx.webpopix.org/userguide/newconnectors/ for more detailed examples
# Download the demo examples here: http://rsmlx.webpopix.org/Rsmlx/Rsmlx10_demos.zip
}
}
