% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getEstimatedResiduals}
\alias{getEstimatedResiduals}
\title{Get estimated residuals}
\usage{
getEstimatedResiduals()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the residuals computed from the individual predictions obtained 
with the estimated individual parameters:
}
\examples{
\dontrun{
# Assume that the Monolix project "warfarinPKPD.mlxtran" has been loaded
r = getEstimatedResiduals()  # r is a list with elements "y1" and "y2" 

# See http://rsmlx.webpopix.org/userguide/newconnectors/ for more detailed examples
# Download the demo examples here: http://rsmlx.webpopix.org/installation
}
}
