% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConnectors.R
\name{getSimulatedPredictions}
\alias{getSimulatedPredictions}
\title{Get simulated predictions}
\usage{
getSimulatedPredictions()
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the simulated individual parameters :
}
\examples{
\dontrun{
# Assume that the Monolix project "warfarinPKPD.mlxtran" has been loaded
r = getSimulatedPredictions()  # r is a list with elements "Cc" and "E" 

# See http://monolix.lixoft.com/rsmlx/newconnectors/ for more detailed examples
# Download the demo examples here: http://monolix.lixoft.com/rsmlx/installation
}
}
