% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiManager.R
\name{initRsmlx}
\alias{initRsmlx}
\title{Initialize Rsmlx library}
\usage{
initRsmlx()
}
\value{
A boolean equaling TRUE if the initialization has been successful and FALSE if not.
}
\description{
Initialize Rsmlx library
}
\examples{
\dontrun{
initRsmlx()
}
}
