% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsmlxTools.R
\name{initRsmlx}
\alias{initRsmlx}
\title{Initialize Rsmlx library}
\usage{
initRsmlx(path = NULL)
}
\value{
A list:
\itemize{
  \item \code{software}: the software that is used (should be monolix with Rsmlx)
  \item \code{path}: the path to MonolixSuite
  \item \code{version}: the version of MonolixSuite that is used
  \item \code{status}: boolean equaling TRUE if the initialization has been successful.
}
}
\description{
Initialize Rsmlx library
}
\examples{
\dontrun{
initRsmlx()  # print the info about Monolix and lixoftConnectors
initRsmlx(path="C:/ProgramData/Lixoft/MonolixSuite2019R1")  # use MonolixSuite 2019R1
}
}
