% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcPlot.R
\name{plotVpc}
\alias{plotVpc}
\title{Plot VPC}
\usage{
plotVpc(vpcData, obsData, obsName, timeName, settings = NULL, theme = NULL)
}
\arguments{
\item{vpcData}{(dataframe) dataframe with bins, vpc empirical and theoretical data.}

\item{obsData}{(dataframe) dataframe with observed data.}

\item{obsName}{(str) Name of observation in dataset header.}

\item{timeName}{(str) Name of time in dataset header.}

\item{settings}{[optional] a list of settings for plot
\itemize{
\item \code{observedData} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove observed data (default FALSE). 
\item \code{censoredData} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove censored data (default FALSE). 
\item \code{empiricalData} (\emph{bool}) (\emph{optional}) Add/remove empirical data (default FALSE). 
empirical percentiles where continuous data /
empirical probability where discrete data /
empirical curve where event data
\item \code{theoreticalData} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove theoreticalData data (default FALSE). 
predicted percentiles where continuous data /
theoreticalData probability where discrete data /
\item \code{predictionInterval} (\emph{bool}) (\emph{optional}) Display Prediction interval (default FALSE)
\item \code{survivalCurve} (\emph{bool}) (\emph{optional}) (\emph{event data}) Add/remove plot for survival function (Kapan-Meier plot) (default TRUE).
\item \code{meanNumberEventsCurve} (\emph{bool}) (\emph{optional}) (\emph{event data}) Add/remove plot for mean number of events per individual (default FALSE).
\item \code{outliersDots} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove red dots indicating empirical percentiles that are outside prediction intervals (default TRUE).
\item \code{outlierAreas} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove red areas indicating empirical percentiles that are outside prediction intervals (default TRUE).
\item \code{legend} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove legend (default FALSE).
\item \code{grid} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove grid (default FALSE).
\item \code{xlogScale} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove log scale for x axis (default FALSE).
\item \code{ylogScale} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) Add/remove log scale for x axis (default FALSE).
\item \code{linearInterpolation} (\emph{bool}) (\emph{optional}) (\emph{continuous data}) If TRUE set piece wise display for prediction intervals, else show bins as rectangular (default TRUE).
\item \code{xlab} (\emph{string}) (\emph{optional}) Time label
(default "time" if timeName = "time", "time since last dose" if timeName = "timeSinceLastDose").
\item \code{ylab} (\emph{string}) (\emph{optional}) y label (default observation name).
\item \code{binLimits} (\emph{bool}) (\emph{optional}) Add/remove vertical lines on the scatter plots to indicate the bins (default FALSE).
}}

\item{theme}{(\emph{vpc_theme}) (\emph{optional}) theme to be used in VPC. Expects list of class vpc_theme created with function createVpcTheme()}
}
\value{
a ggplot2 object
}
\description{
Plot VPC
}
\seealso{
\link{vpcStats} \link{createVpcTheme} \link{plotVPC}
}
