% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcPlot.R
\name{plotVPC}
\alias{plotVPC}
\title{Visual Predictive Check}
\usage{
plotVPC(project, stratifier = NULL, settings = NULL, preferences = NULL)
}
\arguments{
\item{project}{Monolix project}

\item{stratifier}{[optional] a list of settings for to stratify vpc plots:
\itemize{
\item \code{covariateSplit} vector of covariate names used to split vpc plot 
(no split by default).
\item \code{covariateFilter} list of covariate names used to filter vpc plot
names are covariate names and values are category name in case of categorical covariate
and group id in case of continuous covariate.
(no filter by default).
\item \code{covariateScaling} list of continuous covariate scaling
Names are continuous covariate names and values are vector of break doubles.
If not scaling defined, by default break is defined at the median.
}}

\item{settings}{[optional] a list of settings for to stratify vpc plots:
\itemize{
\item \code{grid} boolean, if TRUE, grid is displayed (default TRUE).
\item \code{legend} boolean, if TRUE, legend is displayed (default FALSE).
\item \code{displayObservedData} boolean, if TRUE, observed data is displayed (default FALSE).
\item \code{displayCensoredData} boolean, if TRUE, censored data is displayed (default FALSE).
\item \code{displayEmpiricalPercentiles} boolean, if TRUE, empirical percentiles are displayed (default TRUE).
\item \code{displayPredictedPercentiles} boolean, if TRUE, theoretical percentiles are displayed (default FALSE).
\item \code{level} integer, level for prediction intervals computation (default 90).
\item \code{higherPercentile} integer, higher percentile for empirical and predicted percentiles computation (default 90).
\item \code{outliersDots} boolean, if TRUE, red dots indicating empirical percentiles that are outside prediction intervals are displayed (default TRUE).
\item \code{outlierAreas} boolean, if TRUE, red areas indicating empirical percentiles that are outside prediction intervals are displayed (default TRUE).
\item \code{correctedPredictions} boolean, if TRUE, pcVPC are computed using Uppsala prediction correction (default FALSE).
\item \code{linearInterpolation} boolean, if FALSE, Set piece wise display for prediction intervals (default TRUE).
\item \code{useCensoredData} boolean, Choose to use BLQ data or to ignore it to compute the VPC (default TRUE).
\item \code{censoredData} in {`simulated`, `loq`}, BLQ data can be simulated, or can be equal to the limit of quantification (LOQ) (default `simulated`).
\item \code{binLimits} boolean, if TRUE, vertical lines added to the scatter plots to indicate the bins (default FALSE).
\item \code{binCriteria} Bining criteria in {`manual`, `equalwidth`, `equalsize`, `leastsquare`} (default `leastsquare`).
\item \code{useFixedNbBins} boolean, fixed number of bins. To define if `binCriteria` in {`equalwidth`, `equalsize`, `leastsquare`} (default FALSE).
\item \code{fixedNbBins} integer, fixed number of bins. To define if `useFixedNbBins` is TRUE (default 10).
\item \code{estimatedNbBins} integer vector of size 2, Bin range for automatic selection. To define if `useFixedNbBins` is FALSE (default `c(5, 30)`).
\item \code{nbBinData} integer vector of size 2, range for the number of data points per bin. To define if `binCriteria` in {`equalwidth`, `equalsize`, `leastsquare`} (default `c(10, 200)`).
\item \code{manualBins} vector, Manually define bins boundaries. To define if `binCriteria` is `manual`.
\item \code{xlogScale} boolean, if TRUE, plot in x log scale (default FALSE).
\item \code{ylogScale} boolean, if TRUE, plot in y log scale (default FALSE).
\item \code{xlabel} in {`time`, `timeSinceLastDose`} (default `time`).
}}

\item{preferences}{[optional] boolean to define if plot preferences}
}
\description{
Generate Visual predictive check plots
}
\examples{
\dontrun{
  stratifier <- list(covariateSplit = c("sex", "wt"),
                     covariateFilter = list("age" = 1),
                     covariateScaling = list("age" = c(65), "wt" = c(70)))
  settings <- list(binCriteria = "manual", manualBins = c(0, 24, 48, 72, 96, 120))
  plotVPC(project="RsmlxDemo1.mlxtran", stratifier=stratifier)
  plotVPC(project="RsmlxDemo1.mlxtran", settings=settings)
  plotVPC(project="RsmlxDemo1.mlxtran", settings=list(xlabel = "timeSinceLastDose"))
} 
}
