% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpcPlot.R
\name{plotVpc}
\alias{plotVpc}
\title{Plot VPC}
\usage{
plotVpc(
  vpcData,
  obsData,
  obsName,
  timeName,
  settings = NULL,
  theme = NULL,
  minLog = 0
)
}
\arguments{
\item{vpcData}{(dataframe) dataframe with bins, vpc empirical and theoretical data.}

\item{obsData}{(dataframe) dataframe with observed data.}

\item{obsName}{(str) Name of observation in dataset header.}

\item{timeName}{(str) Name of time in dataset header.}

\item{settings}{[optional] a list of settings for plot
\itemize{
\item \code{observedData} [optional][continuous data] (boolean) Add/remove observed data (default FALSE). 
\item \code{censoredData} [optional][continuous data] (boolean) Add/remove censored data (default FALSE). 
\item \code{empiricalData} [optional] (boolean) Add/remove empirical data (default FALSE). 
empirical percentiles where continuous data /
empirical probability where discrete data /
empirical curve where event data
\item \code{theoreticalData} [optional] (boolean) Add/remove theoreticalData data (default FALSE). 
predicted percentiles where continuous data /
theoreticalData probability where discrete data /
\item \code{predictionInterval} [optional] (boolean) Display Prediction interval (default FALSE)
\item \code{survivalCurve} [optional][event data] (boolean) Add/remove plot for survival function (Kapan-Meier plot) (default TRUE).
\item \code{meanNumberEventsCurve} [optional][event data] (boolean) Add/remove plot for mean number of events per individual (default FALSE).
\item \code{outliersDots} [optional] (boolean) Add/remove red dots indicating empirical percentiles that are outside prediction intervals (default TRUE).
\item \code{outlierAreas} [optional] (boolean) Add/remove red areas indicating empirical percentiles that are outside prediction intervals (default TRUE).
\item \code{legend} [optional] (boolean) Add/remove legend (default FALSE).
\item \code{grid} [optional] (boolean) Add/remove grid (default FALSE).
\item \code{xlogScale} [optional] (boolean) Add/remove log scale for x axis (default FALSE).
\item \code{ylogScale} [optional] (boolean) Add/remove log scale for x axis (default FALSE).
\item \code{linearInterpolation} [optional] (boolean) If TRUE set piece wise display for prediction intervals, else show bins as rectangular (default TRUE).
\item \code{xlab} [optional] (str) Time label (default "time" if time = "time", "time since last dose" if time = "timeSinceLastDose").
\item \code{ylab} [optional] (str) y label (default observation name).
\item \code{binLimits} [optional] (boolean) Add/remove vertical lines on the scatter plots to indicate the bins (default FALSE).
}}

\item{vpcTheme}{[optional] theme to be used in VPC. Expects list of class vpc_theme created with function createVpcTheme()}
}
\value{
a ggplot2 object
}
\description{
Plot VPC
}
\seealso{
\link{vpcStats} \link{createVpcTheme} \link{plotVPC}
}
