\name{RsmlxDemo1.project}
\docType{data}
\alias{RsmlxDemo1.project}
\title{Monolix project for warfarin PK - 1}
\description{
  RsmlxDemo2.mlxtran is a Monolix project for modelling the pharmacokinetics (PK) of warfarin 
using a PK model with parameters ka, V, Cl. There is no covariate in the model.
}
\usage{RsmlxDemo1.project}
\format{A text file}
\source{Monolix project}
\references{Rsmlx doumentation}
\keyword{datasets}

