% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmlx.R
\name{testmlx}
\alias{testmlx}
\title{Statistical tests}
\usage{
testmlx(project, test.list = c("covariate", "randomEffect", "correlation",
  "residual"), plot = FALSE)
}
\arguments{
\item{project}{a Monolix project}

\item{test.list}{the tests to perform}

\item{plot}{FALSSE/TRUE  plot the results}
}
\value{
a list of data frames (one data frame per output).
}
\description{
Get the individual predictions obtained with the simulated individual parameters :
}
\examples{
\dontrun{
r = getSimulatedPredictions()
names(r)
    "Cc"  "E"
}
}
