\name{warfarin.data}
\docType{data}
\alias{warfarin.data}
\title{warfarin PKPD data}
\description{
  The warfarin PK and PD data for 32 patients
}
\usage{warfarin.data}
\format{A csv file}
\source{Monolix demos}
\references{O’Reilly (1968). Studies on coumarin anticoagulant drugs. Initiation of warfarin therapy without a loading dose. Circulation 1968, 38:169-177.}
\keyword{datasets}

